+++ samples/WebApp/Scripts/App.ts
@@ -60,7 +60,7 @@
     });
 
     S.putDataSeriesTurnaround(cli)(dataSeries).then(result => {
-        if (result !== undefined) {
+        if (Gluon.Option.isSome(result)) {
             console.log("putDataSeriesTurnaround => ", result.DataPoints, result.DataPoints.length);
         }
     });
@@ -90,7 +90,7 @@
     dictExample.setAt("two", 2);
     dictExample.setAt("three", 13);
     S.convertDict(cli)(dictExample).then(x => {
-        if (x !== undefined) {
+        if (Gluon.Option.isSome(x)) {
             x.forEach((key, value) => {
                 console.log("dict:", key, value);
             });
@@ -129,9 +129,10 @@
     S.dictCheck(cli)(d0).then(x => console.log("dict check returned ok", x));
 
     S.getTwoDates(cli)().then(pair => {
-        if (pair !== undefined) {
-            console.log("getTwoDates => ", pair[0], pair[1])
-            S.getTwoDatesBack(cli)(pair[0], pair[1]).then(result => console.log("getTwoDatesBack => ", result));
+        if (Gluon.Option.isSome(pair)) {
+            const [x, y] = pair;
+            console.log("getTwoDates => ", x, y)
+            S.getTwoDatesBack(cli)(x, y).then(result => console.log("getTwoDatesBack => ", result));
         }
     });